requirements.txt
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
fake-useragent==1.4.0
frozenlist==1.4.1
future==0.18.3
greenlet==3.0.3
idna==3.6
install==1.3.5
lxml==5.1.0
lzstring==1.0.4
multidict==6.0.5
playwright==1.41.2
psycopg2-binary==2.9.9
pyee==11.0.1
python-dotenv==1.0.1
python-socks==2.4.4
requests==2.31.0
soupsieve==2.5
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
